Pychart/requirements.txt
# Uncomment to enable Azure Monitor OpenTelemetry
# Ref: aka.ms/functions-azure-monitor-python
# azure-monitor-opentelemetry

azure-functions

# Test libraries for plotting and psychrometric calculations
matplotlib
psychrolib
psychrochart
